+++ src/NRImport.cpp
@@ -68,7 +68,7 @@ SEXP emr_import(SEXP _track, SEXP _db_id, SEXP _categorical, SEXP _src, SEXP _ad
                 int curr_db_idx = g_db->get_db_idx(g_db->track_info(trackname)->db_id);
                 int req_db_idx = g_db->get_db_idx(db_id);
 
-                //Do not allowe writing a track with the same name to a db dir
+                //Do not allow writing a track with the same name to a db dir
                 //With lower priority. This will only confuse the user.
                 if (curr_db_idx > req_db_idx) {
                     verror("Can not write track to %s the same track already exists in %s. This kind of write has no effect.", db_id.c_str(), g_db->track_info(trackname)->db_id.c_str());